move longer examples in a separate tests folder
